ai-job-agent/requirements.txt
crewai
crewai-tools
google-search-results
python-dotenv
streamlit
pydantic